Fix: fix the bug which cause the cards to re-render during onchange of name input field.
Change the body of the randomArray method. Now it created a array of
objects. Each object has fruit name and uniqueId.
The cause of the error is we can't directly set uuid on the key in react.
Cause it will set the new uuid on each render which will cause the component
to get update again and again.