Refactor: Move the state from child component to the parent.
Move all the states in MainGame.js to App.js, cause states need to be share between other
components like GameOverPopup.js.
Add a logic to update the bestScore if the current score is higher.
Add a reset method which whill reset the necessary states to their
initial states and let the user continue the game.